Optimize background color drawing.
Before it would color each cell individually. Now it colors up to a row at a
time, for the maximum extent a color is used. Black is never drawn, since we
just cleared the framebuffer to black, and colors are only changed if it makes
sense to do so (so if you draw blue, black, blue, this will only need to set
the color to blue once, if the previous line hadn't already set it to that).

Less state change, less draws.

This lets scrolling keep up on my 4K display when highlighting text as you go.